picturas-bg-remover-tool-ms/pyproject.toml
[tool.poetry]
name = "picturas-bg-remover-tool-ms"
version = "0.1.0"
description = "PictuRAS microservice (tool) that removes the background of an input image."
authors = ["Afonso Costa <d13690@di.uminho.pt>", "Paulo Sousa <d7157@di.uminho.pt>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.12,<3.13"
pika = "^1.3.2"
pydantic = "^2.10.3"
pillow = "^11.0.0"
rembg = "2.0.61"
onnxruntime = "^1.20.1"


[tool.poetry.group.dev.dependencies]
yapf = "^0.43.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
